Display information for each active pool.
Previous implementation selected the highest priority pool to
display.  This commit adds support for displaying information for
each pool.  It displays pool URL, then on the next refresh, displays
stats from that pool - the quota and accepted difficulty.  Then it
does the same for the next pool. Since a single pool can have
multiple ports this removes the stripping of port information and
favors replacement of TLD which is more common.

This includes a few other fixes for the display to even things out a
bit.